ingest/requirements.txt
nhl-api-py==0.6.2
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
